chore: relocated files and moved signing out logic to custom hook
